Solved AoC2021 Day 6 Part 1

+++ aoc2021/aoc2021-kotlin/src/test/kotlin/de/havox_design/aoc2021/day06/Day06Test.kt
@@ -10,25 +10,25 @@ class Day06Test {
 
     @ParameterizedTest
     @MethodSource("getDataForTestProcessPart1")
-    fun testProcessPart1(filename: String, expectedResult: Long) =
+    fun testProcessPart1(filename: String, expectedResult: Int) =
         Laternfish(filename).processPart1().shouldBe(expectedResult)
 
     @ParameterizedTest
     @MethodSource("getDataForTestProcessPart2")
-    fun testProcessPart2(filename: String, expectedResult: Long) =
+    fun testProcessPart2(filename: String, expectedResult: Int) =
         Laternfish(filename).processPart2().shouldBe(expectedResult)
 
     companion object {
         @JvmStatic
         private fun getDataForTestProcessPart1(): Stream<Arguments> =
             Stream.of(
-                Arguments.of("de/havox_design/aoc2021/day06/day06part1sample.txt", 5934L)
+                Arguments.of("de/havox_design/aoc2021/day06/day06part1sample.txt", 5934)
             )
 
         @JvmStatic
         private fun getDataForTestProcessPart2(): Stream<Arguments> =
             Stream.of(
-                Arguments.of("de/havox_design/aoc2021/day06/day06part2sample.txt", 0L)
+                Arguments.of("de/havox_design/aoc2021/day06/day06part2sample.txt", 0)
             )
     }
 }
